Problem: pages are caching and displaying incorrect information
Solution: Add code to cause pages 'not' to cache from : http://www.jguru.com/faq/view.jsp?EID=377

Note the paragraph from microsoft's website:
"It is highly recommended that all Web servers use a scheme for the expiration of all Web pages. It is bad practice for a Web server not to supply expiration information via the HTTP Expires response header for every resource returned to requesting clients. Most browsers and intermediate proxies today respect this expiration information and use it to increase the efficiency of communications over the network."