Starting code for a PHP version of eQuest.  Changes include the following:
The tables web.ResponseSetResp and web.Response are being combined into a single table called web.Response to make it easier to define response sets.

The columns ValidationCode and ValidationMessage in the table web.SurveyQuestion are being replaced with the single column ValidationCheck (like the table web.ChildQuestion).

The automatically generated column PatientID is being replaced by the column SubjectID, and its length is being increased from 10 to 20 characters.