Case 9306
Moved my sql close statements to right after I was done with each statement instead of having them all on the bottom.  I had done this with a couple of the statements but accidentally left a line where I was trying to close the same statement again.  Apparently 
if ($insert_date) {
	sasql_stmt_close($insert_date);
}
Doesn't do what I wanted to since it tried to close a closed statement.