Case 9062:
Removed Home link in header since it was linking to a page the subject didn't have access to in MonthSelection.php, QuestionnaireDE.php, EntrySelection.php, and CalendarDE.php.

Moved check for complete answers for the diagnosis and travel methods for questions 4 and 5 in the questionnaire from StoreQuestionsDE.php to JS in QuestionnaireDE.php.  This eliminated the need for the missing session variable and caf_is_subquestion_missing.