Case 9574:
Hi Siobhan,

Here would be my suggestion for choosing web only. I would put this on the instructions page in the white space to the right of where they choose how to enter their responses - maybe with a box or something around it to make it stand out.

Choose 1:
Submit questionnaires on the web only. Stop receiving paper copies
Continue receiving paper copies of the questionnaires
Stop participating in the Colds and Flu Questionnaire Study

This may change with other suggestions, but just getting something out there gives people something to look at.

Thanks!
Sue
Siobhan: Can we do a message if they check 'stop participating' and ask them if they are sure they want to withdraw from the Colds and Flu Questionnaire Study? 
----
Still need to implement the stop diaries option in MonthSelection.php
----

Added an eligibility check in caf_load_session_variables to prevent subjects from using a bookmark of the calendar or questionnaire to continue a diary when they've withdrawn.