Case 15038
I changed the workflow for adding subject consents.  Now, the user will select the consent document before providing the details for the items in that consent.

I also changed some of the IDs being used in the PHP to make more sense.

Another thing I changed was the way I was dealing with the editing of subject consents.  Before I was deleting and re-inserting the subject document items as they were edited due to the old way we were storing them.  Now that all the items for each consent document is always stored I changed the SQL to edit existing records instead.