Cases 17948 and 17942:
I changed the error message that lets the user know there is a more recent reference year available when linking to a supplement to give the ID that matches the survey date if it exists.

I added a clear fields button so when the user returns to the supplement entry form after submitting, she can clear the fields and enter a new supplement.