Case 17478:
Finished the changes for this branch.

I changed the reasons for no colo checkboxes to be radio buttons since the user shouldn't be able to check more than one.  I looked at the data to make sure it hadn't happen before.

I added the new fields to the record_final_colo procedure.