Case 12271:
Added code to turn off auto-commit in BaseDAOWrapper.java, commit on success and rollback on failure and then turn auto-commit back on.  Most of the Java code uses the wrapper.  A few Java files don't and I have to figure out how to implement this in the JSP files that use Java on the page.