Case 18960:
Fixed a bug where the user could select "Other" without specifying a reason for the colo question on the OFU Procedures page.  Also added a JavaScript check for both the OFU and EOT Procedures page to stop the user from entering a reason for one of the other choices.  Before it would just silently clear out that field.