Case 20107:
Added validation to the date form completed fields on the child measurements, parent measurements, participant contact info, and status change forms to make sure it's not before the informed consent signed date and not in the future.

Fixed a bug in SubmitChildMeasurements.php and SubmitParentMeasurements.php where it was not selecting from the new status history view.

Fixed the case of "RA terminated" in StatusChange.php