Case 20121 Status Change Form allows change from withdrawn to terminated and vice versa
I changed the form to load the withdrawal/termination data if there is any.  The status radio buttons are disabled if the participant is already withdrawn or terminated.

I added a check in dba.record_status_change as well to prevent a status change when the participant is withdrawn or terminated.