Case 20857:
Found and fixed a few more bugs.

Fixed a bug where PHYSSAME was always checked after a constraint violation.

Fixed a bug where the database data was filled into the fields instead of the submitted data after returning from an error.

Fixed an undefined index in the logs for spouse_phone.